Middleware Wallet

+++ fermat-api/src/main/java/com/bitdubai/fermat_api/layer/_12_middleware/wallet/CryptoAccount.java
@@ -6,7 +6,8 @@ import com.bitdubai.fermat_api.layer._1_definition.enums.CryptoCurrency;
  * Created by ciencias on 2/15/15.
  */
 public interface CryptoAccount {
-    public Double getBalance();
+    
+    public long getBalance();
 
     public CryptoCurrency getCryptoCurrency() ;
